deploy-zscaler-mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deploy-zscaler-mcp"
version = "0.1.0"
description = "CLI tool for deploying Zscaler MCP servers with AI assistants"
license = "Apache-2.0"
requires-python = ">=3.11"
readme = "README.md"
authors = [
    {name = "Your Name"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Telecommunications Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Networking",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "fastmcp>=2.5.1",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[project.urls]
Homepage = "https://github.com/your-username/deploy-zscaler-mcp"
Repository = "https://github.com/your-username/deploy-zscaler-mcp"
Issues = "https://github.com/your-username/deploy-zscaler-mcp/issues"

[project.scripts]
deploy-zscaler-mcp = "deploy_zscaler_mcp.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]